# openapi.yaml
openapi: 3.1.0
info:
  title: OKR Tracking System API
  description: |
    ## OKR Tracking System — Garant Bank

    REST API for managing **Objectives & Key Results (OKR)**, organizational structure, user management, and multi-source performance evaluations.

    ### Key Features
    - **OKR Management** — Create and track Objectives with weighted Key Results
    - **Multi-Source Evaluations** — Director (star rating), HR (letter grade), and Business Block assessments
    - **Score Calculation** — Automatic weighted scoring: 60% OKR + 20% Director + 20% HR
    - **Organizational Hierarchy** — Divisions → Departments → Objectives → Key Results
    - **Role-Based Access Control** — 6 roles with fine-grained permissions
    - **Excel Export** — Export all OKR data to Excel spreadsheets

    ### Authentication
    All endpoints (except login and export) require a **JWT Bearer token**.
    1. Call `POST /api/auth/login` with your credentials
    2. Copy the `token` from the response
    3. Click the **Authorize** button above and enter the token

    ### Roles
    | Role | Description |
    |------|-------------|
    | `ADMIN` | Full system access, user management |
    | `DIRECTOR` | Division oversight, star-based evaluations |
    | `DEPARTMENT_LEADER` | Department OKR management |
    | `HR` | Letter-grade evaluations (A/B/C/D) |
    | `BUSINESS_BLOCK` | Business block star-based evaluations |
    | `EMPLOYEE` | View own profile and assigned departments |
  version: 1.0.0
...
servers:
...
security:
- Bearer Authentication: []
...
paths:
...
  /api/objectives/{id}:
    put:
      tags:
      - Objectives
      summary: Update objective
      description: Update an objective's name and weight. Requires edit permission
        for the parent department.
      operationId: updateObjective
      parameters:
      - name: id
        in: path
        description: Objective ID
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ObjectiveDTO"
        required: true
      responses:
        "200":
          description: Objective updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ObjectiveDTO"
        "403":
          description: No edit permission
        "404":
          description: Objective not found
    delete:
      tags:
      - Objectives
      summary: Delete objective
      description: Delete an objective and all its key results. Requires edit permission
        for the parent department.
      operationId: deleteObjective
      parameters:
      - name: id
        in: path
        description: Objective ID
        required: true
        schema:
          type: string
      responses:
        "204":
          description: Objective deleted
        "403":
          description: No edit permission
        "404":
          description: Objective not found
  /api/key-results/{id}:
    put:
      tags:
      - Key Results
      summary: Update key result
      description: "Update a key result's properties including name, description,\
        \ metric type, thresholds, and weight."
      operationId: updateKeyResult
      parameters:
      - name: id
        in: path
        description: Key Result ID
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KeyResultDTO"
        required: true
      responses:
        "200":
          description: Key result updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KeyResultDTO"
        "403":
          description: No edit permission
        "404":
          description: Key result not found
    delete:
      tags:
      - Key Results
      summary: Delete key result
      description: Permanently delete a key result. Requires edit permission for the
        parent department.
      operationId: deleteKeyResult
      parameters:
      - name: id
        in: path
        description: Key Result ID
        required: true
        schema:
          type: string
      responses:
        "204":
          description: Key result deleted
        "403":
          description: No edit permission
        "404":
          description: Key result not found
  /api/key-results/{id}/actual-value:
    put:
      tags:
      - Key Results
      summary: Update key result actual value
      description: Update only the actual/measured value of a key result. The score
        is automatically recalculated based on the new value and the configured thresholds.
      operationId: updateKeyResultActualValue
      parameters:
      - name: id
        in: path
        description: Key Result ID
        required: true
        schema:
          type: string
      requestBody:
        description: Map with key `actualValue` containing the new measured value
        content:
          application/json:
            schema:
              type: string
              example:
                actualValue: "95.5"
        required: true
      responses:
        "200":
          description: Actual value updated and score recalculated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KeyResultDTO"
        "403":
          description: No edit permission
        "404":
          description: Key result not found
...
  /api/departments/{id}:
    get:
      tags:
      - Departments
      summary: Get department by ID
      description: "Returns a single department with its objectives, key results,\
        \ and computed scores."
      operationId: getDepartment
      parameters:
      - name: id
        in: path
        description: Department ID
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Department found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepartmentDTO"
        "404":
          description: Department not found
    put:
      tags:
      - Departments
      summary: Update department
      description: "Update a department's name and other properties. Requires edit\
        \ permission for the department (ADMIN, DIRECTOR, or assigned DEPARTMENT_LEADER)."
      operationId: updateDepartment
      parameters:
      - name: id
        in: path
        description: Department ID
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentDTO"
        required: true
      responses:
        "200":
          description: Department updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepartmentDTO"
        "403":
          description: No edit permission for this department
        "404":
          description: Department not found
    delete:
      tags:
      - Departments
      summary: Delete department
      description: Permanently delete a department and all its objectives/key results.
        **Requires ADMIN role.**
      operationId: deleteDepartment
      parameters:
      - name: id
        in: path
        description: Department ID
        required: true
        schema:
          type: string
      responses:
        "204":
          description: Department deleted
        "403":
          description: Only ADMIN can delete departments
        "404":
          description: Department not found
...
  /api/objectives/{objectiveId}/key-results:
    post:
      tags:
      - Key Results
      summary: Create key result
      description: "Create a new key result within an objective. Configure metric\
        \ type (HIGHER_BETTER, LOWER_BETTER, QUALITATIVE), unit, weight, and threshold\
        \ values. Key result weights within an objective should sum to 100%."
      operationId: createKeyResult
      parameters:
      - name: objectiveId
        in: path
        description: Objective ID
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KeyResultDTO"
        required: true
      responses:
        "200":
          description: Key result created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KeyResultDTO"
        "403":
          description: No edit permission
        "404":
          description: Objective not found
...
  /api/departments:
    get:
      tags:
      - Departments
      summary: Get all departments
      description: "Returns all departments with their objectives, key results, and\
        \ computed scores. Accessible to all authenticated users."
      operationId: getAllDepartments
      responses:
        "200":
          description: List of all departments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DepartmentDTO"
        "401":
          description: Not authenticated
    post:
      tags:
      - Departments
      summary: Create department
      description: Create a new department. **Requires ADMIN or DIRECTOR role.** The
        department must be assigned to an existing division via the `divisionId` field.
      operationId: createDepartment
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentDTO"
        required: true
      responses:
        "200":
          description: Department created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepartmentDTO"
        "403":
          description: Insufficient permissions
  /api/departments/{departmentId}/objectives:
    post:
      tags:
      - Objectives
      summary: Create objective
      description: Create a new objective within a department. The objective weight
        should be between 0-100 and all objective weights in a department should sum
        to 100%. Requires edit permission for the department.
      operationId: createObjective
      parameters:
      - name: departmentId
        in: path
        description: Department ID
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ObjectiveDTO"
        required: true
      responses:
        "200":
          description: Objective created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ObjectiveDTO"
        "403":
          description: No edit permission for this department
        "404":
          description: Department not found
  /api/demo/load:
    post:
      tags:
      - Demo Data
      summary: Load demo data
      description: "Load sample OKR data for testing and demonstration purposes. Creates\
        \ sample divisions, departments, objectives, and key results. **Requires ADMIN\
        \ role.**"
      operationId: loadDemoData
      responses:
        "200":
          description: Demo data loaded — returns all departments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DepartmentDTO"
        "403":
          description: Only ADMIN can load demo data
...
  /api/export/excel:
    get:
      tags:
      - Export
      summary: Export OKRs to Excel
      description: "Export all departments, objectives, and key results to an Excel\
        \ (.xlsx) file. The file includes scores and status information. **Public\
        \ endpoint — no authentication required.**"
      operationId: exportToExcel
      responses:
        "200":
          description: Excel file generated
          content:
            application/octet-stream: {}
        "500":
          description: Export failed
      security: []
...
  /api/departments/{id}/scores:
    get:
      tags:
      - Departments
      summary: Get department scores with evaluations
      description: "Returns detailed score breakdown for a department including automatic\
        \ OKR score (60%), Director evaluation (20%), HR evaluation (20%), Business\
        \ Block assessment, and the combined final score."
      operationId: getDepartmentScores
      parameters:
      - name: id
        in: path
        description: Department ID
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Department score details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepartmentScoreResult"
        "404":
          description: Department not found
  /api/auth/me:
    get:
      tags:
      - Authentication
      summary: Get current user
      description: Returns the profile information of the currently authenticated
        user based on the JWT token.
      operationId: getCurrentUser
      responses:
        "200":
          description: Current user details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDTO"
        "401":
          description: Not authenticated
...
components:
  schemas:
...
    ObjectiveDTO:
      type: object
      description: Objective with key results and computed score
      properties:
        id:
          type: string
          description: Objective ID
          example: obj-001
        name:
          type: string
          description: Objective name
          example: Improve System Reliability
        weight:
          type: integer
          format: int32
          description: Weight as percentage of department total (0-100). All objectives
            in a department should sum to 100.
          example: 40
        departmentId:
          type: string
          description: Parent department ID
          example: dept-001
        keyResults:
          type: array
          description: Key results under this objective
          items:
            $ref: "#/components/schemas/KeyResultDTO"
        score:
          $ref: "#/components/schemas/ScoreResult"
          description: Computed score based on key result achievements
...
    DepartmentDTO:
      type: object
      description: "Department with objectives, scores, and division info"
      properties:
        division:
          $ref: "#/components/schemas/DivisionSummaryDTO"
          description: Parent division summary
        divisionId:
          type: string
          description: Division ID (used when creating a department)
          example: div-001
        id:
          type: string
          description: Department ID
          example: dept-001
        name:
          type: string
          description: Department name
          example: Software Development
        objectives:
          type: array
          description: List of objectives in this department
          items:
            $ref: "#/components/schemas/ObjectiveDTO"
        score:
          $ref: "#/components/schemas/ScoreResult"
          description: Automatic OKR score computed from key results
        finalScore:
          $ref: "#/components/schemas/ScoreResult"
          description: "Final combined score: 60% OKR + 20% Director + 20% HR"
        hasAllEvaluations:
          type: boolean
          description: True if the department has both Director and HR evaluations
          example: true
...
  securitySchemes:
    Bearer Authentication:
      type: http
      description: Enter the JWT token obtained from `POST /api/auth/login`
      name: Bearer Authentication
      scheme: bearer
      bearerFormat: JWT